Undo  Templated comparators Op::[isGt,isGE,isEq,isLE,isLT]

+++ fuzzylite/src/imex/FisImporter.cpp
@@ -309,7 +309,7 @@ namespace fl {
     std::string FisImporter::translateProposition(scalar code, Variable* variable) const {
         int intPart = (int) std::floor(std::fabs(code)) - 1;
         scalar fracPart = std::fmod(std::fabs(code), 1.0);
-        if (intPart >= variable->numberOfTerms()) {
+        if (intPart >= static_cast<int>(variable->numberOfTerms())) {
             std::ostringstream ex;
             ex << "[syntax error] the code <" << code << "> refers to a term "
                     "out of range from variable <" << variable->getName() << ">";
